assistant_tools: Rename `RegexSearchTool` module to match the others
