+++ crates/assistant_tools/src/regex_search_tool.rs
